app/requirements.txt
fastapi == 0.110.0
sqlalchemy == 2.0.29
psycopg2 == 2.9.9
python-dotenv == 1.0.1
loguru == 0.7.2
uvicorn == 0.29.0
pika == 1.3.2
python-jose == 3.3.0
python-multipart== 0.0.9
streamlit == 1.33.0
pytest == 8.1.1
httpx == 0.27.0
passlib == 1.7.4
xgboost == 2.1.0
joblib == 1.4.0